Improve mobile reading experience
Hi, I read your work on my phone a lot and the current layout is
restricted to 65em, which results in a lot of horizontal
scrolling. This commit changes it so that the max-width is still 65em,
but it will also scale to smaller devices. After making this change, I
also noticed that the navigation h1 tag looked a bit weird on smaller
sizes, so I moved the margin from the h1 to the navigation wrapping
.header element.

I tested the main page and a blog post on desktop chrome, android,
ipad pro, and iphone, but didn't check every post so there could be
code somewhere that "breaks out" of the layout and requires horizontal
scolling.